+++ opentok.model.ts
@@ -1,17 +1,3 @@
-export enum Signal {
-  CloseSession = 'closeSession',
-  StageRequest = 'stageRequest',
-  StageRequestApproved = 'stageRequestApproved',
-  StageRequestDenied = 'stageRequestDenied',
-  StageRequestCanceled = 'stageRequestCanceled',
-  StageRevoked = 'stageRevoked',
-  HandLowered = 'handLowered',
-  HandRaised = 'handRaised',
-  VideoShared = 'videoShared',
-  VideoActive = 'videoActive',
-  VideoSync = 'videoSync',
-}
-
 export enum OtEventNames {
   AudioBlocked = 'audioBlocked',
   AudioLevelUpdated = 'audioLevelUpdated',
